+++ lib/main.dart
@@ -1,4 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex_app/app.dart';
+import 'package:pokedex_app/presentation/app.dart';
 
 void main() => runApp(const PokedexApp());
