+++ lib/presentation/app.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex_app/features/home/presentation/home_page.dart';
-import 'package:pokedex_app/resources/strings.dart';
-import 'package:pokedex_app/resources/theme.dart';
+import 'package:pokedex_app/presentation/features/pokedex/pokedex_page.dart';
+import 'package:pokedex_app/presentation/resources/strings.dart';
+import 'package:pokedex_app/presentation/resources/theme.dart';
 
 class PokedexApp extends StatelessWidget {
   const PokedexApp({super.key});
@@ -13,7 +13,7 @@ class PokedexApp extends StatelessWidget {
       themeMode: ThemeMode.system, //TODO: create logic for theme swap
       theme: PokedexTheme.light,
       darkTheme: PokedexTheme.dark,
-      home: const HomePage(),
+      home: const PokedexPage(),
     );
   }
 }
