+++ lib/presentation/features/pokedex/pokedex_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex_app/resources/strings.dart';
+import 'package:pokedex_app/presentation/resources/strings.dart';
 
-class HomePage extends StatelessWidget {
-  const HomePage({super.key});
+class PokedexPage extends StatelessWidget {
+  const PokedexPage({super.key});
 
   @override
   Widget build(BuildContext context) {
